api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
email-validator==2.0.0.post2
passlib==1.7.4
bcrypt==4.0.1  # Required for password hashing
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.0
stripe==6.5.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.0.1
PyJWT==2.8.0  # Changed from jwt to PyJWT which is more standard
jinja2==3.1.2  # Required for test client templates